OMS Adapter: isolate viewability helpers to avoid eager side effects
